Represent References
We want to index maps with objects. Unfortunately the javascript Map
type cannot use object keys as keys are compared by referential
equality. However, OrderedMap (from `immutable`) does have this
property, if the keys are `ValueObject`s. `Record`s are the relevant
types: `Record(…)` creates a type constructor, with a default value, and
that constructor creates the `Reference` values.

The centre of the crossword is at (0,0), with the coordinate system
extending right with increasing x coordinate and up with increasing y.
The `centre` value itself is only used to create a `Reference` type,
using `typeof`.