Represent grids as OrderedMap of CellProps
These CellMaps represent cell positions as a mapping from `Reference`s
to `CellProps`.

The display itself uses css grid, meaning the divs corresponding to the
cells are in a flat, not a nested, structure.

It would be good to represent the grid using a Store Comonad but in
practice that might be best as an api on top of this OrderedMap.