Add functionality to display grids
What we actually display is not what we represent. We default to an
empty grid, of the supplied size. This means that we can represent
larger grids in the internal state, but only display the required
sub-part. The state can also be just the required updates to this empty
grid — and does not in fact have to be an OrderedMap.