Create generators for the serialization
Crossword grids are serialized by spiralling out from the centre. This
ensures the requisite 180° symmetry (as we only store n+1 of the 2n+1
cells) and that changing the size of the grid preserves the
serialization.

These generators produce streams of values which help with this process.
It is probably _not_ the best way to do this but the relevant
simplifications can come later.