Fix BigInt pow bug
It is possible to use the `**` operator with `BigInt`s, but not
`Math.pow`. Unfortunately the build is is outputting `Math.pow`, which
_seems to be due to the typescript conversion — but

1.  it’s only for builds — running the app (which uses babel) is not
affected, and
2.  it has been fixed? https://github.com/microsoft/TypeScript/issues/30398 so maybe it’s a regression.

More likely is there’s an issue with my setup but it was easier (in this
case) to convert the booleans to numbers and run the exponentiation on
those. I could even actually use the booleans to return `-1` or `1` as
appropriate but given I want to replace the whole thing I’m a bit 🤷.