Represent lights in state, not grid
The lights state is a map of `Reference`s to `boolean`s, not to
`CellProps`. The latter is the state of the grid, not the grid layout.

The numbering algorithm depends on the lights, and the grid display
depends on the numbering, so the state should flow in the one direction
from lights ⇒ numbering ⇒ display. Therefore I should store the lights
separately to data generated from the lights.

As a result of this change, some identifiers have been renamed more
appropriately.