Add functionality to track transits
A transit is a clue crossing a cell. In a valid grid, there is either 1
or 2 clues crossing each cell.

To calculate transits, we only need to consider the down transit in the
immediately upper cell and the across transit in the cell immediately to
the left. These are generated as we progress through the algorithm,
meaning we do need to traverse the grid in order. However, it is a less
strict order than for clue starts — we technically need to traverse
sequential diagonals.